requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.3
distlib==0.3.1
Django==3.1.6
djangorestframework==3.12.4
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
# psycopg2==2.8.6
pytz==2021.1
requests==2.26.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.4.2
whitenoise==6.2.0
